moving cache logic to rewards service and removing unnecessary joins
