+++ src/modules/rewards/adapter/db/op-reward-fixture.ts
@@ -27,7 +27,6 @@ export function mockOpRewardEntity(overrides: Partial<OpReward>) {
     amount: "1000000000000000000",
     amountUsd: "1",
     rewardTokenId: 1,
-    rewardTokenPriceId: 1,
     depositDate: new Date(),
     ...overrides,
   };
