+++ src/modules/rewards/services/arb-rebate-service.ts
@@ -1,8 +1,7 @@
-import { CACHE_MANAGER, Inject, Injectable, Logger } from "@nestjs/common";
+import { Injectable, Logger } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { DataSource, In, Repository } from "typeorm";
 import BigNumber from "bignumber.js";
-import { Cache } from "cache-manager";
 import { ethers } from "ethers";
 import { DateTime } from "luxon";
 
@@ -17,8 +16,6 @@ import { ArbReward } from "../model/arb-reward.entity";
 import { GetRewardsQuery } from "../entrypoints/http/dto";
 
 const ARB_REBATE_RATE = 0.95;
-const getArbEarnedRewardsCacheKey = (address: string) => `arbRewards:earned:${address}`;
-const getArbRebatesSummaryCacheKey = (address: string) => `arbRebates:summary:${address}`;
 
 type PartialDeposit = Pick<
   Deposit,
@@ -41,7 +38,6 @@ export class ArbRebateService {
   constructor(
     @InjectRepository(Deposit) readonly depositRepository: Repository<Deposit>,
     @InjectRepository(ArbReward) readonly arbRewardRepository: Repository<ArbReward>,
-    @Inject(CACHE_MANAGER) private cacheManager: Cache,
     private marketPriceService: MarketPriceService,
     private ethProvidersService: EthProvidersService,
     private appConfig: AppConfig,
@@ -51,30 +47,17 @@ export class ArbRebateService {
   public async getEarnedRewards(userAddress: string) {
     userAddress = assertValidAddress(userAddress);
 
-    const data = await this.cacheManager.get(getArbEarnedRewardsCacheKey(userAddress));
-    if (data) return data;
-
     const baseQuery = this.buildBaseQuery(this.arbRewardRepository.createQueryBuilder("r"), userAddress);
     const { arbRewards } = await baseQuery
       .select("SUM(CAST(r.amount as DECIMAL))", "arbRewards")
+      .where("r.isClaimed = :isClaimed", { isClaimed: true })
       .getRawOne<{ arbRewards: string }>();
-
-    if (this.appConfig.values.app.cacheDuration.rebatesData) {
-      await this.cacheManager.set(
-        getArbEarnedRewardsCacheKey(userAddress),
-        arbRewards,
-        this.appConfig.values.app.cacheDuration.rebatesData,
-      );
-    }
 
     return arbRewards;
   }
 
   public async getArbRebatesSummary(userAddress: string) {
     userAddress = assertValidAddress(userAddress);
-
-    let data = await this.cacheManager.get(getArbRebatesSummaryCacheKey(userAddress));
-    if (data) return data;
 
     const baseQuery = this.buildBaseQuery(this.arbRewardRepository.createQueryBuilder("r"), userAddress);
     baseQuery.andWhere("r.isClaimed = :isClaimed", { isClaimed: false });
@@ -95,22 +78,12 @@ export class ArbRebateService {
         }>(),
     ]);
 
-    data = {
+    return {
       depositsCount: parseInt(depositsCount),
       unclaimedRewards,
       volumeUsd,
       claimableRewards: "0",
     };
-
-    if (this.appConfig.values.app.cacheDuration.rebatesData) {
-      await this.cacheManager.set(
-        getArbRebatesSummaryCacheKey(userAddress),
-        data,
-        this.appConfig.values.app.cacheDuration.rebatesData,
-      );
-    }
-
-    return data;
   }
 
   public async getArbRebateRewards(query: GetRewardsQuery) {
@@ -120,11 +93,7 @@ export class ArbRebateService {
 
     const baseQuery = this.buildBaseQuery(this.arbRewardRepository.createQueryBuilder("r"), userAddress);
 
-    const rewardsQuery = baseQuery
-      .leftJoinAndSelect("r.rewardToken", "rewardToken")
-      .orderBy("r.depositDate", "DESC")
-      .limit(limit)
-      .offset(offset);
+    const rewardsQuery = baseQuery.orderBy("r.depositDate", "DESC").limit(limit).offset(offset);
     const [rewards, total] = await rewardsQuery.getManyAndCount();
 
     const depositPrimaryKeys = rewards.map((reward) => reward.depositPrimaryKey);
